+++ plugin/src/net/groboclown/idea/p4ic/changes/P4ChangeList.java
@@ -37,6 +37,7 @@
 package net.groboclown.idea.p4ic.changes;
 
 import net.groboclown.idea.p4ic.server.P4FileInfo;
+import net.groboclown.idea.p4ic.server.P4Job;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -53,12 +54,12 @@ public class P4ChangeList {
     private final Set<P4FileInfo> files;
     private final String comment;
     private final String owner;
-    private final List<String> jobIds;
+    private final List<P4Job> jobIds;
     private final Date lastUpdateTime = new Date();
 
     public P4ChangeList(@NotNull final P4ChangeListId id, @NotNull final Collection<P4FileInfo> files,
                         @Nullable final String comment, @Nullable final String owner,
-                        @Nullable Collection<String> jobIds) {
+                        @Nullable Collection<P4Job> jobIds) {
         this.id = id;
         this.files = Collections.unmodifiableSet(new HashSet<P4FileInfo>(files));
         this.comment = comment;
@@ -66,7 +67,7 @@ public class P4ChangeList {
         if (jobIds == null || jobIds.isEmpty()) {
             this.jobIds = Collections.emptyList();
         } else {
-            this.jobIds = Collections.unmodifiableList(new ArrayList<String>(jobIds));
+            this.jobIds = Collections.unmodifiableList(new ArrayList<P4Job>(jobIds));
         }
     }
 
@@ -96,7 +97,12 @@ public class P4ChangeList {
     }
 
     @NotNull
-    public List<String> getJobIdToStatus() {
+    public List<P4Job> getJobs() {
         return jobIds;
+    }
+
+    @Override
+    public String toString() {
+        return "[" + id + ": jobs " + jobIds + ", files " + files + "]";
     }
 }